Created PreviewGenerator module
