+++ composer.json
@@ -75,7 +75,8 @@
             "Shlinkio\\Shlink\\Core\\": "module/Core/src",
             "Shlinkio\\Shlink\\Common\\": "module/Common/src",
             "Shlinkio\\Shlink\\EventDispatcher\\": "module/EventDispatcher/src",
-            "Shlinkio\\Shlink\\IpGeolocation\\": "module/IpGeolocation/src/"
+            "Shlinkio\\Shlink\\IpGeolocation\\": "module/IpGeolocation/src/",
+            "Shlinkio\\Shlink\\PreviewGenerator\\": "module/PreviewGenerator/src/"
         },
         "files": [
             "module/Common/functions/functions.php",
@@ -94,6 +95,7 @@
             "ShlinkioTest\\Shlink\\Common\\": "module/Common/test",
             "ShlinkioTest\\Shlink\\EventDispatcher\\": "module/EventDispatcher/test",
             "ShlinkioTest\\Shlink\\IpGeolocation\\": "module/IpGeolocation/test",
+            "ShlinkioTest\\Shlink\\PreviewGenerator\\": "module/PreviewGenerator/test",
             "Shlinkio\\Shlink\\TestUtils\\": "module/TestUtils/src"
         }
     },
