+++ module/PreviewGenerator/src/Service/PreviewGeneratorInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Common\Service;
+namespace Shlinkio\Shlink\PreviewGenerator\Service;
 
 use Shlinkio\Shlink\Common\Exception\PreviewGenerationException;
